authorized crud endpoints

+++ routes/packages.js
@@ -6,9 +6,10 @@ const Package = require('../models/Package');
 const Test = require('../models/Test');
 const runner = require('../runner');
 
-const jwtMiddleware = require('express-jwt');
-
 dotenv.config();
+
+const jwt = require('express-jwt');
+const jwtMiddleware = jwt({ secret: process.env.JWT_SECRET, algorithms: ['HS256'] });
 
 router
   .get('/', async (req, res) => {
@@ -27,7 +28,15 @@ router
     const test = await Test.findOne({ where: { packageId: req.params.id, id: req.params.testId }});
     res.send(test);
   })
-  .post('/', jwtMiddleware({ secret: process.env.JWT_SECRET, algorithms: ['HS256'] }), async (req, res) => {
+  .post('/', jwtMiddleware, async (req, res) => {
+    if (!req.user.isAdmin) {
+      res.send({
+        severity: 'ERROR',
+        messages: ['Csak admin tud csomagot létrehozni!']
+      });
+      return;
+    }
+
     try {
       await Package.create(req.body);
     } catch(err) {
@@ -44,7 +53,15 @@ router
     const results = await runner(tests, req.body.url);
     res.send(results);
   })
-  .put('/', jwtMiddleware({ secret: process.env.JWT_SECRET, algorithms: ['HS256'] }), async (req, res) => {
+  .put('/', jwtMiddleware, async (req, res) => {
+    if (!req.user.isAdmin) {
+      res.send({
+        severity: 'ERROR',
+        messages: ['Csak admin tud csomagot módosítani!']
+      });
+      return;
+    }
+
     try {
       await Package.update(req.body, { where: { id: req.body.id } });
     } catch(err) {
@@ -56,7 +73,15 @@ router
     }
     res.sendStatus(200);
   })
-  .delete('/', jwtMiddleware({ secret: process.env.JWT_SECRET, algorithms: ['HS256'] }), async (req, res) => {
+  .delete('/', jwtMiddleware, async (req, res) => {
+    if (!req.user.isAdmin) {
+      res.send({
+        severity: 'ERROR',
+        messages: ['Csak admin tud csomagot törölni!']
+      });
+      return;
+    }
+
     try {
       await Package.destroy({ where: { id: req.body.id}});
     } catch(err) {
